feat(employee): integrate Pagination component into Employees page
