+++ frontend/src/components/Pagination.jsx
@@ -1,89 +1,108 @@
 import {
+  ChevronsLeft,
   ChevronLeft,
   ChevronRight,
-  ChevronsLeft,
   ChevronsRight,
 } from "lucide-react";
 
-const Pagination = () => {
+const Pagination = ({ page, setPage, limit, setLimit, meta }) => {
+  const handleRowsChange = (e) => {
+    setLimit(Number(e.target.value));
+    setPage(1);
+  };
+
+  const renderPageButtons = () => {
+    const pages = [];
+    const totalPages = meta.totalPages || 1;
+    const maxButtons = 5;
+    const start = Math.max(1, page - 2);
+    const end = Math.min(totalPages, start + maxButtons - 1);
+
+    if (start > 1) {
+      pages.push(1);
+      if (start > 2) pages.push("...");
+    }
+
+    for (let i = start; i <= end; i++) pages.push(i);
+
+    if (end < totalPages) {
+      if (end < totalPages - 1) pages.push("...");
+      pages.push(totalPages);
+    }
+
+    return pages.map((num, i) =>
+      num === "..." ? (
+        <span key={i} className="px-3 text-cyan-500">
+          ...
+        </span>
+      ) : (
+        <button
+          key={i}
+          onClick={() => setPage(num)}
+          className={`px-3 py-1 rounded text-sm ${
+            num === page
+              ? "bg-cyan-600 text-white font-semibold"
+              : "hover:bg-cyan-50 text-cyan-800"
+          }`}
+        >
+          {num}
+        </button>
+      )
+    );
+  };
+
   return (
-    // Pagination Toolbar
-    <div className="bg-gray-50 px-4 py-3 border-t border-gray-200 sm:px-6">
-      <div className="flex flex-col sm:flex-row items-center justify-between gap-4">
-        {/* Rows per page */}
-        <div className="flex items-center gap-2">
-          <span className="text-sm text-gray-700">Rows per page:</span>
-          <select
-            // value={pageSize}
-            // onChange={(e) => setPageSize(Number(e.target.value))}
-            className="border border-gray-300 rounded px-2 py-1 text-sm focus:ring-2 focus:ring-blue-500"
-          >
-            <option value={10}>10</option>
-            <option value={20}>20</option>
-            <option value={50}>50</option>
-          </select>
-        </div>
+    <div className="bg-white mt-4 rounded-xl flex items-center justify-center px-4 py-3 ">
+      {/* Left section — Rows per page */}
+      <div className="flex items-center gap-2 mr-auto text-sm text-cyan-700">
+        <span>Rows per page:</span>
+        <select value={limit} onChange={handleRowsChange}>
+          {[5, 10, 15, 20, 25, 50].map((row) => (
+            <option key={row} value={row}>
+              {row}
+            </option>
+          ))}
+        </select>
+        <span className="ml-4">
+          {meta.total > 0
+            ? `${meta.from}-${meta.to} of ${meta.total}`
+            : "No records"}
+        </span>
+      </div>
 
-        {/* Range text */}
-        <div className="text-sm text-gray-700">
-          {/* {start}–{end} of {total} */}
-        </div>
+      {/* Center section — Page numbers */}
+      <div className="flex items-center gap-1">
+        <button
+          onClick={() => setPage(1)}
+          disabled={!meta.hasPrev}
+          className="p-2 rounded hover:bg-cyan-100 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          <ChevronsLeft className="w-4 h-4 text-cyan-700" />
+        </button>
+        <button
+          onClick={() => setPage((prev) => Math.max(prev - 1, 1))}
+          disabled={!meta.hasPrev}
+          className="p-2 rounded hover:bg-cyan-100 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          <ChevronLeft className="w-4 h-4 text-cyan-700" />
+        </button>
 
-        {/* Page navigation */}
-        <div className="flex items-center gap-1">
-          <button
-            // onClick={() => setPage(1)}
-            // disabled={page === 1}
-            aria-label="First page"
-            className="p-2 rounded hover:bg-gray-200 disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            <ChevronsLeft className="w-4 h-4" />
-          </button>
-          <button
-            // onClick={() => setPage(page - 1)}
-            // disabled={page === 1}
-            aria-label="Previous page"
-            className="p-2 rounded hover:bg-gray-200 disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            <ChevronLeft className="w-4 h-4" />
-          </button>
+        {renderPageButtons()}
 
-          {/* {getPageNumbers().map((pageNum, idx) => (
-            pageNum === '...' ? (
-              <span key={`ellipsis-${idx}`} className="px-3 py-1">...</span>
-            ) : (
-              <button
-                key={pageNum}
-                onClick={() => setPage(pageNum)}
-                aria-current={page === pageNum ? 'page' : undefined}
-                className={`px-3 py-1 rounded text-sm ${
-                  page === pageNum
-                    ? 'bg-blue-600 text-white font-medium'
-                    : 'hover:bg-gray-200 text-gray-700'
-                }`}
-              >
-                {pageNum}
-              </button>
-            )
-          ))} */}
-
-          <button
-            // onClick={() => setPage(page + 1)}
-            // disabled={page === totalPages || totalPages === 0}
-            aria-label="Next page"
-            className="p-2 rounded hover:bg-gray-200 disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            <ChevronRight className="w-4 h-4" />
-          </button>
-          <button
-            // onClick={() => setPage(totalPages)}
-            // disabled={page === totalPages || totalPages === 0}
-            aria-label="Last page"
-            className="p-2 rounded hover:bg-gray-200 disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            <ChevronsRight className="w-4 h-4" />
-          </button>
-        </div>
+        <button
+          onClick={() => setPage((prev) => Math.min(prev + 1, meta.totalPages))}
+          disabled={!meta.hasNext}
+          className="p-2 rounded hover:bg-cyan-100 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          <ChevronRight className="w-4 h-4 text-cyan-700" />
+        </button>
+        <button
+          onClick={() => setPage(meta.totalPages)}
+          disabled={!meta.hasNext}
+          className="p-2 rounded hover:bg-cyan-100 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          <ChevronsRight className="w-4 h-4 text-cyan-700" />
+        </button>
       </div>
     </div>
   );
